addons: share testcfg.py
